"Updated GarnishModel to use double for cantidad, added garnishs list to GarnishViewModel, and modified ProductsClassViewModel to load garnishes and update GarnishViewModel."

+++ lib/displays/restaurant/models/garnish_model.dart
@@ -11,7 +11,7 @@ class GarnishModel {
   int? fProducto;
   int? fUnidadMedida;
   int? fBodega;
-  int? cantidad;
+  double? cantidad;
   String desProducto;
   String desUnidadMedida;
   String? fDesProducto;
